server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography==42.0.0",
    "email-validator>=2.3.0",
    "fastapi==0.109.0",
    "openai>=1.0.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dateutil==2.8.2",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "uvicorn[standard]==0.27.0",
]
